fix: исправить все замечания
