+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -4,24 +4,20 @@ import java.util.Collection;
 import java.util.List;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Positive;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpMethod;
 import org.springframework.web.bind.annotation.*;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.FilmService;
 
+@RequiredArgsConstructor
 @RestController
 @RequestMapping("/films")
 @Slf4j
 public class FilmController {
 
     private final FilmService filmService;
-
-    @Autowired
-    public FilmController(FilmService filmService) {
-        this.filmService = filmService;
-    }
 
     @PostMapping
     public Film postFilm(@Valid @RequestBody Film film) {
